[logical-types] Split logical.rs into multiple files
