+++ datafusion/common/src/scalar/logical/logical_map.rs
@@ -1,0 +1,23 @@
+use std::cmp::Ordering;
+use crate::types::{logical_map, LogicalFieldRef, LogicalTypeRef};
+
+/// TODO logical-types
+#[derive(Clone, PartialEq, Eq, Hash, Debug)]
+pub struct LogicalMap {
+    value_type: LogicalFieldRef,
+    // TODO logical-types
+}
+
+impl LogicalMap {
+    /// Returns the [`LogicalTypeRef`] for [`self`].
+    pub fn logical_type(&self) -> LogicalTypeRef {
+        logical_map(self.value_type.clone())
+    }
+}
+
+
+impl PartialOrd for LogicalMap {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        todo!("logical-types")
+    }
+}
